requirements.txt
opencv-python-headless~=4.5.1.48
streamlit~=0.75.0
numpy~=1.19.5
Pillow~=8.1.0
tensorflow~=2.4.1
